tests/wpt/tests/css/css-text-decor/parsing/text-decoration-skip-spaces-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Decoration Test: getComputedStyle().textDecorationSkipSpaces</title>
<link rel="help" href="https://drafts.csswg.org/css-text-decor-4/#text-decoration-skip-spaces-property">
<meta name="assert" content="text-decoration-skip-spaces computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("text-decoration-skip-spaces", "none");
test_computed_value("text-decoration-skip-spaces", "all");
test_computed_value("text-decoration-skip-spaces", "start");
test_computed_value("text-decoration-skip-spaces", "end");
test_computed_value("text-decoration-skip-spaces", "start end");
</script>
</body>
</html>
